+++ eveuniverse/tests/test_models.py
@@ -10,6 +10,11 @@ from django.utils.timezone import now
 
 from .my_test_data import EsiClientStub, BravadoOperationStub
 from ..helpers import meters_to_ly
+from ..constants import (
+    EVE_CATEGORY_ID_BLUEPRINT,
+    EVE_CATEGORY_ID_SKIN,
+    EVE_CATEGORY_ID_STRUCTURE,
+)
 from ..models import (
     EveUniverseBaseModel,
     EsiMapping,
@@ -1090,7 +1095,7 @@ class TestEveType(NoSocketsTestCase):
     @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
     @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
     def test_can_create_icon_url_1(self, mock_esi):
-        """icon from regular type"""
+        """icon from regular type, automatically detected"""
         mock_esi.client = EsiClientStub()
 
         eve_type, created = EveType.objects.get_or_create_esi(id=603)
@@ -1102,7 +1107,7 @@ class TestEveType(NoSocketsTestCase):
     @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
     @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
     def test_can_create_icon_url_2(self, mock_esi):
-        """icon from blueprint type"""
+        """icon from blueprint type, automatically detected"""
         mock_esi.client = EsiClientStub()
 
         eve_type, created = EveType.objects.get_or_create_esi(id=950)
@@ -1114,7 +1119,7 @@ class TestEveType(NoSocketsTestCase):
     @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
     @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
     def test_can_create_icon_url_3(self, mock_esi):
-        """icon from regular type"""
+        """icon from regular type, preset as blueprint"""
         mock_esi.client = EsiClientStub()
 
         eve_type, created = EveType.objects.get_or_create_esi(id=603)
@@ -1126,15 +1131,88 @@ class TestEveType(NoSocketsTestCase):
 
     @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
     @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
-    def test_can_create_icon_url_4(self, mock_esi):
-        """icon from blueprint type"""
+    def test_can_create_icon_url_3a(self, mock_esi):
+        """icon from regular type, preset as blueprint"""
+        mock_esi.client = EsiClientStub()
+
+        eve_type, created = EveType.objects.get_or_create_esi(id=603)
+        self.assertTrue(created)
+        self.assertEqual(
+            eve_type.icon_url(size=256, category_id=EVE_CATEGORY_ID_BLUEPRINT),
+            "https://images.evetech.net/types/603/bp?size=256",
+        )
+
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
+    def test_can_create_icon_url_5(self, mock_esi):
+        """when called for SKIN type, will return dummy SKIN URL with requested size"""
+        mock_esi.client = EsiClientStub()
+
+        eve_type, created = EveType.objects.get_or_create_esi(id=34599)
+        self.assertTrue(created)
+        self.assertIn("skin_generic_64.png", eve_type.icon_url(size=64))
+
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
+    def test_can_create_icon_url_5a(self, mock_esi):
+        """when called for SKIN type, will return dummy SKIN URL with requested size"""
+        mock_esi.client = EsiClientStub()
+
+        eve_type, created = EveType.objects.get_or_create_esi(id=34599)
+        self.assertTrue(created)
+        self.assertIn("skin_generic_32.png", eve_type.icon_url(size=32))
+
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
+    def test_can_create_icon_url_5b(self, mock_esi):
+        """when called for SKIN type, will return dummy SKIN URL with requested size"""
+        mock_esi.client = EsiClientStub()
+
+        eve_type, created = EveType.objects.get_or_create_esi(id=34599)
+        self.assertTrue(created)
+        self.assertIn("skin_generic_128.png", eve_type.icon_url(size=128))
+
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
+    def test_can_create_icon_url_5c(self, mock_esi):
+        """when called for SKIN type and size is invalid, then raise exception"""
+        mock_esi.client = EsiClientStub()
+
+        eve_type, created = EveType.objects.get_or_create_esi(id=34599)
+        self.assertTrue(created)
+        with self.assertRaises(ValueError):
+            eve_type.icon_url(size=512)
+
+        with self.assertRaises(ValueError):
+            eve_type.icon_url(size=1024)
+
+        with self.assertRaises(ValueError):
+            eve_type.icon_url(size=31)
+
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
+    def test_can_create_icon_url_6(self, mock_esi):
+        """when called for non SKIN type and SKIN is forced, then return SKIN URL"""
         mock_esi.client = EsiClientStub()
 
         eve_type, created = EveType.objects.get_or_create_esi(id=950)
         self.assertTrue(created)
+        self.assertIn(
+            "skin_generic_128.png",
+            eve_type.icon_url(size=128, category_id=EVE_CATEGORY_ID_SKIN),
+        )
+
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)
+    @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_DOGMAS", False)
+    def test_can_create_icon_url_7(self, mock_esi):
+        """when called for SKIN type and regular is forced, then return regular URL"""
+        mock_esi.client = EsiClientStub()
+
+        eve_type, created = EveType.objects.get_or_create_esi(id=34599)
+        self.assertTrue(created)
         self.assertEqual(
-            eve_type.icon_url(size=256, is_blueprint=False),
-            "https://images.evetech.net/types/950/icon?size=256",
+            eve_type.icon_url(size=256, category_id=EVE_CATEGORY_ID_STRUCTURE),
+            "https://images.evetech.net/types/34599/icon?size=256",
         )
 
     @patch(MODULE_PATH + ".EVEUNIVERSE_LOAD_MARKET_GROUPS", False)